// src/models/IoK8sApiCoreV1PersistentVolumeSpec/types.ts
export interface IoK8sApiCoreV1PersistentVolumeSpecModelOptions {
  accessModes?: Array<string>;
  awsElasticBlockStore?: IoK8sApiCoreV1AWSElasticBlockStoreVolumeSource;
  azureDisk?: IoK8sApiCoreV1AzureDiskVolumeSource;
  azureFile?: IoK8sApiCoreV1AzureFilePersistentVolumeSource;
  capacity?: { [key: string]: string };
  cephfs?: IoK8sApiCoreV1CephFSPersistentVolumeSource;
  cinder?: IoK8sApiCoreV1CinderPersistentVolumeSource;
  claimRef?: IoK8sApiCoreV1ObjectReference;
  csi?: IoK8sApiCoreV1CSIPersistentVolumeSource;
  fc?: IoK8sApiCoreV1FCVolumeSource;
  flexVolume?: IoK8sApiCoreV1FlexPersistentVolumeSource;
  flocker?: IoK8sApiCoreV1FlockerVolumeSource;
  gcePersistentDisk?: IoK8sApiCoreV1GCEPersistentDiskVolumeSource;
  glusterfs?: IoK8sApiCoreV1GlusterfsPersistentVolumeSource;
  hostPath?: IoK8sApiCoreV1HostPathVolumeSource;
  iscsi?: IoK8sApiCoreV1ISCSIPersistentVolumeSource;
  local?: IoK8sApiCoreV1LocalVolumeSource;
  mountOptions?: Array<string>;
  nfs?: IoK8sApiCoreV1NFSVolumeSource;
  nodeAffinity?: IoK8sApiCoreV1VolumeNodeAffinity;
  persistentVolumeReclaimPolicy?: string;
  photonPersistentDisk?: IoK8sApiCoreV1PhotonPersistentDiskVolumeSource;
  portworxVolume?: IoK8sApiCoreV1PortworxVolumeSource;
  quobyte?: IoK8sApiCoreV1QuobyteVolumeSource;
  rbd?: IoK8sApiCoreV1RBDPersistentVolumeSource;
  scaleIO?: IoK8sApiCoreV1ScaleIOPersistentVolumeSource;
  storageClassName?: string;
  storageos?: IoK8sApiCoreV1StorageOSPersistentVolumeSource;
  volumeAttributesClassName?: string;
  volumeMode?: string;
  vsphereVolume?: IoK8sApiCoreV1VsphereVirtualDiskVolumeSource;
}